refactor: Extract test_utils.py and improve OCaml 5.3.0 GC workaround
- Create shared test_utils.py module with common functions
- More aggressive GC settings for aarch64/ppc64le under QEMU:
  s=128M, H=256M, o=200, d=1 (single domain)
- Add test_utils.py to recipe.yaml test files
