+++ recipe/testing/test-opam-activation.py
@@ -7,43 +7,10 @@ should already be set. The opam root is pre-initialized at build time.
 """
 
 import os
-import platform
 import subprocess
 import sys
 
-
-def get_ocaml_version():
-    """Get OCaml version string."""
-    try:
-        result = subprocess.run(
-            ["ocaml", "-version"],
-            capture_output=True,
-            text=True,
-            check=False,
-        )
-        for word in result.stdout.split():
-            if word[0].isdigit():
-                return word
-    except FileNotFoundError:
-        pass
-    return "unknown"
-
-
-def apply_ocaml_530_workaround():
-    """Apply OCaml 5.3.0 aarch64/ppc64le GC workaround if needed."""
-    ocaml_version = get_ocaml_version()
-    arch = platform.machine().lower()
-
-    print(f"OCaml version: {ocaml_version}")
-    print(f"Architecture: {arch}")
-
-    if ocaml_version.startswith("5.3.") and arch in ("aarch64", "ppc64le", "arm64"):
-        # OCaml 5.3.0 has heap corruption issues on aarch64/ppc64le under QEMU.
-        gc_params = "s=128M,H=256M,o=200"
-        print(f"Applying OCaml 5.3.0 GC workaround ({gc_params})")
-        os.environ["OCAMLRUNPARAM"] = gc_params
-
-    print(f"OCAMLRUNPARAM: {os.environ.get('OCAMLRUNPARAM', '<default>')}")
+from test_utils import apply_ocaml_530_workaround, get_opam_root
 
 
 def check_env(name, expected):
@@ -75,12 +42,7 @@ def main():
         print("[FAIL] CONDA_PREFIX not set - not running in conda environment")
         return 1
 
-    # Platform-specific paths
-    # Windows: conda packages install to Library\share, not share
-    if platform.system() == "Windows":
-        expected_root = os.path.join(conda_prefix, "Library", "share", "opam")
-    else:
-        expected_root = os.path.join(conda_prefix, "share", "opam")
+    expected_root = get_opam_root()
 
     print("\n--- Test: environment variables ---")
     print(f"CONDA_PREFIX: {conda_prefix}")
